data_collection_platform/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
graphene-django==3.2.3
graphene==3.4.3
PyMySQL==1.1.2
mysqlclient==2.2.7 # optionnel si tu préfères le driver natif (nécessite build tools)
requests==2.32.5
python-dotenv==1.2.1
django-cors-headers==4.9.0

gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==3.0.1
psycopg-binary==3.2.12
psycopg==3.2.12
psycopg[binary]==3.2.12